# perf(core-database-postgres): drop redundant and add missing index on blocks