--- a/packages/core-database-postgres/lib/migrations/20181204100000-blocks-generator_public_key-index.sql
+++ b/packages/core-database-postgres/lib/migrations/20181204100000-blocks-generator_public_key-index.sql
@@ -1,0 +1,2 @@
+DROP INDEX "blocks_unique";
+CREATE INDEX IF NOT EXISTS "blocks_generator_public_key" ON ${schema~}.blocks ("generator_public_key");